WebUI: factor out some parts of Home.js to separate components
